// src/api/generated/models/FloorPlan.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Floor plan response model
 */
export type FloorPlan = {
    name: string;
    location_id?: (string | null);
    layout: Record<string, any>;
    is_active?: boolean;
    id: string;
    business_id: string;
    created_at: string;
    updated_at: string;
};