// src/api/generated/models/FloorPlanUpdate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Update floor plan
 */
export type FloorPlanUpdate = {
    name?: (string | null);
    location_id?: (string | null);
    layout?: (Record<string, any> | null);
    is_active?: (boolean | null);
};